-[LOCATIONSERVICE] The location identifier as a long because assets can be used as locations and they have a Long identifier while real locations have a Integer identifier. -[DM] Removed unused classes that are obsolete.

+++ NeoCom.DataManagement/src/main/java-adapter/org/dimensinfin/eveonline/neocom/adapters/ESIDataAdapter.java
@@ -176,7 +176,7 @@ public class ESIDataAdapter {
 		return Futures.immediateFuture(new MarketDataSet(itemId, side));
 	}
 
-	public EsiLocation searchLocation4Id( final Integer locationId ) {
+	public EsiLocation searchLocation4Id( final Long locationId ) {
 		return this.locationCatalogService.searchLocation4Id(locationId);
 	}
